+++ hallo_eltern_cli/commands/stdoutcommand.py
@@ -13,6 +13,9 @@ class StdoutCommand(EmailCommand):
 
         self._process_all = args.process_all
 
+        self._separator = args.separator
+        self._processed_count = 0
+
     @classmethod
     def register_subparser(cls, subparsers):
         parser = register_command_class(
@@ -22,5 +25,13 @@ class StdoutCommand(EmailCommand):
                             action='store_true',
                             help='process all (even already seen) messages')
 
+        parser.add_argument('--separator',
+                            help='separator to print between messages')
+
     def process_email(self, email):
+        if self._processed_count:
+            print(self._separator)
+
         print(email)
+
+        self._processed_count += 1